+++ CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved all API files into the api source set
 - Moved lots of mod logic into the api
 - Updated README
+- Moved Plugin annotation handling to its own class
 
 ### Removed
 - Harvestcraft jar, we fetch from jitpack now